requirements.txt
Flask==3.0.2
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Jinja2==3.1.6
SQLAlchemy==2.0.28
python-dotenv==1.0.1
PyMySQL==1.1.0
cryptography==44.0.3
Werkzeug==3.0.1 